+++ src/main/scala/tigerjython/execute/TigerJythonExecutor.scala
@@ -77,6 +77,7 @@ class TigerJythonExecutor(val id: Int, override val controller: ExecutionControl
           onResult.setError(result)
         else
           onResult.setResult(result)
+
       })
     })